lock: one request at time
fix skip on busy

+++ server.py
@@ -16,30 +16,30 @@ class GenerationServer(BaseHTTPRequestHandler):
 
     def do_POST(self):
         if self.path.startswith('/generate'):
-            free = self.lock.acquire(blocking=False)
-            if not free:
+            if self.lock.acquire(blocking=False):
+                try:
+                    self.send_response(200)
+
+                    content_length = int(self.headers['Content-Length'])
+                    data = self.rfile.read(content_length).decode('utf-8')
+                    args = copy.deepcopy(startup_args)
+                    args.prompt = data
+
+                    pipeline = generator.stable_diffusion_pipeline(args)
+                    img_paths = generator.stable_diffusion_inference(pipeline)
+
+                    f = open(img_paths[0], 'rb')
+                    self.send_header("Content-type", "image/png")
+                    self.end_headers()
+                    self.wfile.write(f.read())
+                    f.close()
+
+                    for img_path in img_paths:
+                        os.remove(img_path)
+                finally:
+                    self.lock.release()
+            else:
                 self.send_error(503, 'Busy')
-            try:
-                self.send_response(200)
-
-                content_length = int(self.headers['Content-Length'])
-                data = self.rfile.read(content_length).decode('utf-8')
-                args = copy.deepcopy(startup_args)
-                args.prompt = data
-
-                pipeline = generator.stable_diffusion_pipeline(args)
-                img_paths = generator.stable_diffusion_inference(pipeline)
-
-                f = open(img_paths[0], 'rb')
-                self.send_header("Content-type", "image/png")
-                self.end_headers()
-                self.wfile.write(f.read())
-                f.close()
-
-                for img_path in img_paths:
-                    os.remove(img_path)
-            finally:
-                self.lock.release()
 
         else:
             self.send_error(404, 'Not found')
